Make `EdgeServers` fields optional
Because `EdgeServers` is modeled from the info server, each field should
be optional in order to loosen the IO coupling between the info-client
and info-server, avoiding crashes caused by missing or deprecated
server-lists on the info-server.

This also loosens the requirement to include all servers when providing
options to the info-client.