fixup! Implement Disklet synchronization
This function seems reasonable. However, this should not be public API. Instead, we want to export a `makeSyncedDisklet(parentPath: Disklet) => Disklet & { sync() => Promise<string[]> }`.

You do `const myDisklet = makeSyncedDisklet(...)`, and then you can do `myDisklet.getText`, `myDisklet.delete`, `myDisklet.setText`, and so forth. This library automatically routes the changes to the right places (/data, /changes, /deleted). Finally, you can do `myDisklet.sync()` to sync with the server.

This keeps the details fully black-boxed. The contents of the "status.json", the layout of the sub-folders (/data, /changes, /deleted), and the details of the sync algorithm are fully abstracted. Remember when we built the change server? We built the "legacy" API compatible with the existing core, and we also defined the "new" API with faster sync. We never implemented the client side of the "new" API, but if we wanted to do that, we would want edge-sync-client to be "black-boxed" so the change can be seamless. If the core has to "know" the format of the "status.json" or the general on-disk layout of the repo, we have failed.

Addresses: https://github.com/EdgeApp/edge-sync-client/pull/19#discussion_r2608550317